pyproject.toml
[project]
name = "factorio-working-group"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<=3.12"
dependencies = [
    "factorio-draftsman>=2.0.1",
    "numpy>=2.2.4",
    "scikit-learn>=1.6.1",
    "tensorboard>=2.19.0",
    "torch>=2.6.0",
    "torch-tb-profiler>=0.4.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]